Remove unnecessary wrapping of private key in encrypted JSON wallet object
In the web3dart library you can get all the same "wallet" feature from
the EthPrivateKey class that you get from the Wallet class. The Wallet
class is a bit of a confusing name, because what it really represents in
a specific way of encoding keys in encrypted form for storage and
sharing.