Publish version 0.3.0
The commit history will look a bit funky because we rolled back v0.2.1
but can't actually remove it, so we pushed a new patch level based on
old commit. This version 0.3.0 is essentially what used to be 0.2.1, but
now is a minor version bump so people will have to be more explicit in
their pubspec to grab it.