Remove Mumbai and replace with Amoy (#42)
This is not a backwards compatible change. After some thought, and discussion, it was decided that you can't simply call something mumbai but point it at Amoy, because Amoy breaks too many things. None of the contract addresses are the same, so devs will need to make a conscious migration to Amoy.

* Update network config for Amoy specific contract addresses and chain config

* Rename all code references from mumbai to amoy

* Update README docs to be pointed at Amoy rather than mumbai. Included migration strategy in the changelog, since our user base isn't huge. No reason to clutter README with migration details. 

* Update included example API token to be pointed at Amoy